Move id-generator to data-structures namespace
