+++ extensions/amp-accordion/1.0/component.js
@@ -22,7 +22,7 @@ import {omit} from '#core/types/object';
 import {
   randomIdGenerator,
   sequentialIdGenerator,
-} from '#core/math/id-generator';
+} from '#core/data-structures/id-generator';
 import {
   useCallback,
   useContext,
