+++ extensions/amp-lightbox-gallery/1.0/consumer.js
@@ -16,7 +16,7 @@
 
 import * as Preact from '#preact';
 import {LightboxGalleryContext} from './context';
-import {sequentialIdGenerator} from '#core/math/id-generator';
+import {sequentialIdGenerator} from '#core/data-structures/id-generator';
 import {useContext, useLayoutEffect, useMemo, useState} from '#preact';
 
 const generateLightboxItemKey = sequentialIdGenerator();
