+++ src/core/data-structures/id-generator.js
